+++ dev/src/router.js
@@ -8,7 +8,6 @@ import Button from '../pages/Button.vue'
 import ButtonGroup from '../pages/ButtonGroup.vue'
 import Card from '../pages/Card.vue'
 import Expandable from '../pages/Expandable.vue'
-import Input from '../pages/Input.vue'
 import Modal from '../pages/Modal.vue'
 import Pill from '../pages/Pill.vue'
 import Select from '../pages/Select.vue'
@@ -18,6 +17,7 @@ import Switch from '../pages/Switch.vue'
 import Tabs from '../pages/Tabs.vue'
 import Tag from '../pages/Tag.vue'
 import Textarea from '../pages/Textarea.vue'
+import Textfield from '../pages/Textfield.vue'
 import Toggle from '../pages/Toggle.vue'
 // import Forms from '../pages/Forms.vue'
 
@@ -31,7 +31,6 @@ export const routes = [
   { path: '/button-group', component: ButtonGroup, name: 'button-group' },
   { path: '/card', component: Card, name: 'card' },
   { path: '/expandable', component: Expandable, name: 'expandable' },
-  { path: '/input', component: Input, name: 'input' },
   { path: '/modal', component: Modal, name: 'modal' },
   { path: '/pill', component: Pill, name: 'pill' },
   { path: '/select', component: Select, name: 'select' },
@@ -41,6 +40,7 @@ export const routes = [
   { path: '/tabs', component: Tabs, name: 'tabs' },
   { path: '/tag', component: Tag, name: 'tag' },
   { path: '/textarea', component: Textarea, name: 'textarea' },
+  { path: '/textfield', component: Textfield, name: 'textfield' },
   { path: '/toggle', component: Toggle, name: 'toggle' },
   // { path: '/forms', component: Forms, name: 'forms' },
 ]
